+++ src/components/App/filters/formatMoney.js
@@ -1,0 +1,5 @@
+import { globalize } from './filters'
+
+export function formatMoney (value) {
+  return globalize('formats.money', { value })
+}