+++ CHANGELOG.md
@@ -10,11 +10,26 @@ for further information about branching and tagging conventions.
 ## [Unreleased]
 #### Added
 - Form validation of doc types in limits form
+- new filter formatMoney
 
 #### Fixed
+- A bug with localization
+- A bug with wrong displaying int64 in AssetAmountFormatter
 - A bug with displaying current issuer public key in AssetManager
 
+#### Removed
+- formatAssetAmount, formatFiatAmount helpers
+- AssetAmountFormatter
+
 ## [1.13.0-rc.0] - 2020-03-25
+#### Fixed
+- A bug with wrong displaying initial preissued amount
+  in SaleRequestManager.DetailsTab
+
+#### Removed
+- util localize
+
+## [1.12.1-rc.0] - 2020-03-25
 #### Fixed
 - Details in createChangeRoleRequest at unblock user
 - A bug with wrong error if exceeded the maximum issuance amount of the asset
@@ -554,8 +569,7 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/admin-panel/compare/1.13.0-rc.0...HEAD
-[1.13.0-rc.0]: https://github.com/tokend/admin-panel/compare/1.12.0...1.13.0-rc.0
+[Unreleased]: https://github.com/tokend/admin-panel/compare/1.12.0...HEAD
 [1.12.0]: https://github.com/tokend/admin-panel/compare/1.12.0-rc.4...1.12.0
 [1.12.0-rc.4]: https://github.com/tokend/admin-panel/compare/1.12.0-rc.3...1.12.0-rc.4
 [1.12.0-rc.3]: https://github.com/tokend/admin-panel/compare/1.12.0-rc.2...1.12.0-rc.3